# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroHub API
  description: 'API operacional de AeroHub: aprovisionamiento de tenants, seguimiento
    de vuelos y distribucion de FIDS en tiempo real (RF-T02, OpenAPI 3.1).'
  contact:
    name: AeroHub Platform Team
  version: 0.1.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /puertas/asignaciones:
    post:
      tags:
      - puertas
      summary: Asignar Puerta Endpoint
      description: Asigna una puerta a un vuelo (RF-O02, PN-05).
      operationId: asignar_puerta_endpoint_puertas_asignaciones_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarPuertaRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsignarPuertaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /puertas/asignaciones/{asignacion_id}/cancelar:
    post:
      tags:
      - puertas
      summary: Cancelar Asignacion Endpoint
      description: 'Cancela una asignacion, liberando la puerta (baja logica, P5:
        nunca

        un DELETE de motor sobre la fila).'
      operationId: cancelar_asignacion_endpoint_puertas_asignaciones__asignacion_id__cancelar_post
      parameters:
      - name: asignacion_id
        in: path
        required: true
        schema:
          type: integer
          title: Asignacion Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /puertas/terminales:
    get:
      tags:
      - puertas
      summary: Listar Terminales Endpoint
      operationId: listar_terminales_endpoint_puertas_terminales_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TerminalListadoResponse'
                type: array
                title: Response Listar Terminales Endpoint Puertas Terminales Get
    post:
      tags:
      - puertas
      summary: Crear Terminal Endpoint
      description: 'Alta de terminal (Fase 3 de docs/diseno/PLAN_CORRECCION_MODULOS.md,

        causa raiz backend: M3 no tenia forma de crear terminales, solo se

        veian si un seed las insertaba a mano).'
      operationId: crear_terminal_endpoint_puertas_terminales_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalCrearRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalCrearResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /puertas:
    post:
      tags:
      - puertas
      summary: Crear Puerta Endpoint
      description: 'Alta de puerta (Fase 3 de docs/diseno/PLAN_CORRECCION_MODULOS.md,

        causa raiz backend: el tablero solo listaba puertas ya existentes,

        ninguna forma de crearlas desde la API).'
      operationId: crear_puerta_endpoint_puertas_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PuertaCrearRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PuertaCrearResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /puertas/{puerta_id}:
    patch:
      tags:
      - puertas
      summary: Editar Puerta Endpoint
      description: 'Edicion de puerta -- terminal, codigo, tipo, envergadura y pasarela

        (Fase 3, mismo hallazgo que el alta: no existia ninguna forma de

        editar una puerta ya creada).'
      operationId: editar_puerta_endpoint_puertas__puerta_id__patch
      parameters:
      - name: puerta_id
        in: path
        required: true
        schema:
          type: integer
          title: Puerta Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PuertaEditarRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /puertas/informes/simple:
    get:
      tags:
      - puertas
      summary: Informe Asignaciones Simple Endpoint
      operationId: informe_asignaciones_simple_endpoint_puertas_informes_simple_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: puerta_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Puerta Id
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /puertas/informes/compuesto:
    get:
      tags:
      - puertas
      summary: Informe Asignaciones Compuesto Endpoint
      operationId: informe_asignaciones_compuesto_endpoint_puertas_informes_compuesto_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AsignarPuertaRequest:
      properties:
        vuelo_id:
          type: integer
          title: Vuelo Id
        puerta_id:
          type: integer
          title: Puerta Id
        inicio_previsto:
          type: string
          format: date-time
          title: Inicio Previsto
        fin_previsto:
          type: string
          format: date-time
          title: Fin Previsto
      type: object
      required:
      - vuelo_id
      - puerta_id
      - inicio_previsto
      - fin_previsto
      title: AsignarPuertaRequest
    AsignarPuertaResponse:
      properties:
        asignacion_id:
          type: string
          title: Asignacion Id
      type: object
      required:
      - asignacion_id
      title: AsignarPuertaResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PuertaCrearRequest:
      properties:
        terminal_id:
          type: integer
          title: Terminal Id
        codigo:
          type: string
          maxLength: 10
          minLength: 1
          title: Codigo
        tipo:
          type: string
          title: Tipo
        envergadura_max_m:
          type: number
          title: Envergadura Max M
        tiene_pasarela:
          type: boolean
          title: Tiene Pasarela
          default: false
      type: object
      required:
      - terminal_id
      - codigo
      - tipo
      - envergadura_max_m
      title: PuertaCrearRequest
    PuertaCrearResponse:
      properties:
        puerta_id:
          type: string
          title: Puerta Id
      type: object
      required:
      - puerta_id
      title: PuertaCrearResponse
    PuertaEditarRequest:
      properties:
        terminal_id:
          type: integer
          title: Terminal Id
        codigo:
          type: string
          maxLength: 10
          minLength: 1
          title: Codigo
        tipo:
          type: string
          title: Tipo
        envergadura_max_m:
          type: number
          title: Envergadura Max M
        tiene_pasarela:
          type: boolean
          title: Tiene Pasarela
          default: false
      type: object
      required:
      - terminal_id
      - codigo
      - tipo
      - envergadura_max_m
      title: PuertaEditarRequest
...
    TerminalCrearRequest:
      properties:
        codigo:
          type: string
          maxLength: 10
          minLength: 1
          title: Codigo
        nombre:
          type: string
          maxLength: 100
          minLength: 1
          title: Nombre
      type: object
      required:
      - codigo
      - nombre
      title: TerminalCrearRequest
    TerminalCrearResponse:
      properties:
        terminal_id:
          type: string
          title: Terminal Id
      type: object
      required:
      - terminal_id
      title: TerminalCrearResponse
    TerminalListadoResponse:
      properties:
        id:
          type: string
          title: Id
        codigo:
          type: string
          title: Codigo
        nombre:
          type: string
          title: Nombre
      type: object
      required:
      - id
      - codigo
      - nombre
      title: TerminalListadoResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...